Add support for storing containers at an arbitrary location
This adds two config parameters, "container_depot" and "container_rootfs,"
that specify where individual containers and the base, read-only filesystem
should live, respectively.

Test plan:
- Ran the specs
- Ran `bundle exec rake setup[/tmp/warden/rootfs]` and then started
  the warden using the modified linux config. Manually created, modified,
  and destroyed containers using the repl.

+++ warden/root/insecure/clear.sh
@@ -5,12 +5,19 @@ set -o errexit
 shopt -s nullglob
 cd $(dirname "${0}")
 
-for instance in instances/*; do
+if [ $# -ne 1 ]; then
+  echo "Usage: ${0} <instances_path>"
+  exit 1
+fi
+
+instances_path=${1}
+
+for instance in instances_path/*; do
   [ -f ${instance}/stop.sh ] && ${instance}/stop.sh || true
 done
 
 sleep 0.1
 
-for instance in instances/*; do
+for instance in instances_path/*; do
   rm -rf ${instance}
 done
